+++ packages/sui-bundler/webpack.config.dev.js
@@ -55,7 +55,7 @@ const webpackConfig = {
   cache: {
     type: 'filesystem',
     cacheDirectory,
-    compression: 'brotli'
+    compression: 'gzip'
   },
   target: 'web',
   optimization: {